split combat into subfolders, update imports project-wide, and document the refactor in progress log.
