+++ combat/servants/kitik_servant.py
@@ -1,5 +1,5 @@
-from combat.passive_triggers import HOOK_ON_LOW_HP, HOOK_ON_TURN_START, register_hook
-from combat.servant_base import ServantBase
+from combat.systems.passive_triggers import HOOK_ON_LOW_HP, HOOK_ON_TURN_START, register_hook
+from combat.entities.servant_base import ServantBase
 
 _HOOKS_REGISTERED = False
 
@@ -92,3 +92,4 @@ def _register_kitik_hooks():
     register_hook(HOOK_ON_TURN_START, "Kitik", on_turn_start)
     register_hook(HOOK_ON_LOW_HP, "Kitik", on_low_hp)
     _HOOKS_REGISTERED = True
+
