+++ combat/ui/combat_renderer.py
@@ -1,9 +1,9 @@
 import pygame
 
-from combat.combat_constants import COMBAT_BG, COMBAT_GOLD
-from combat.combat_log import draw_combat_log
-from combat.combat_ui import draw_action_bar, draw_combat_tracker, draw_player_status, draw_skill_submenu
-from combat.enemy_panel import draw_enemy_panel
+from combat.data.combat_constants import COMBAT_BG, COMBAT_GOLD
+from combat.ui.combat_log import draw_combat_log
+from combat.ui.combat_ui import draw_action_bar, draw_combat_tracker, draw_player_status, draw_skill_submenu
+from combat.ui.enemy_panel import draw_enemy_panel
 
 _FONTS = {}
 
@@ -26,3 +26,4 @@ def render_combat_frame(screen, state, title: str):
     draw_player_status(screen, state)
     draw_action_bar(screen, state)
     draw_skill_submenu(screen, state)
+
