+++ game.log
@@ -700,3 +700,5 @@ cannot load library 'libcairo-2.dll': error 0x7e.  Additionally, ctypes.util.fin
 2026-04-02 17:54:34,283 | INFO | Game closed
 2026-04-02 18:02:02,751 | INFO | Game started
 2026-04-02 18:06:09,088 | INFO | Game closed
+2026-04-03 20:32:13,069 | INFO | Game started
+2026-04-03 20:33:24,489 | INFO | Game closed
